+++ postcss.config.cjs
@@ -1,0 +1,9 @@
+const postcssNesting = require("postcss-nesting");
+const autoprefixer = require("autoprefixer");
+
+module.exports = {
+    plugins: [
+        postcssNesting,
+        autoprefixer,
+    ],
+};