+++ src/components/Breadcrumbs/Breadcrumbs/Breadcrumbs.tsx
@@ -1,16 +1,21 @@
+// FRAMEWORK ===========================================================================================================
 import React from "react";
 
+// FICTOAN =============================================================================================================
 import { Element } from "../../Element";
 
-import { CommonAndHTMLProps } from "../../Element/constants";
-
+// STYLES ==============================================================================================================
 import "./breadcrumbs.css";
+
+// TYPES ===============================================================================================================
+import { CommonAndHTMLProps } from "../../Element/constants";
 
 export type BreadcrumbsElementType = HTMLDivElement;
 export interface BreadcrumbsProps extends CommonAndHTMLProps<BreadcrumbsElementType> {
     children : React.ReactNode;
 }
 
+// COMPONENT ///////////////////////////////////////////////////////////////////////////////////////////////////////////
 export const Breadcrumbs = React.forwardRef<HTMLDivElement, BreadcrumbsProps>(
     ({ children, ...props }, ref) => {
         return (