+++ src/components/Form/Checkbox/Switch.tsx
@@ -1,10 +1,15 @@
+// FRAMEWORK ===========================================================================================================
 import React from "react";
 
+// FICTOAN =============================================================================================================
 import { Element } from "../../Element/Element";
 import { BaseInputComponent } from "../BaseInputComponent/BaseInputComponent";
-import { BaseInputComponentProps } from "../BaseInputComponent/constants";
 
+// STYLES ==============================================================================================================
 import "./switch.css";
+
+// TYPES ===============================================================================================================
+import { BaseInputComponentProps } from "../BaseInputComponent/constants";
 
 export interface SwitchCustomProps {
     size ? : "small" | "medium" | "large";
@@ -14,6 +19,7 @@ export type SwitchElementType = HTMLInputElement;
 export type SwitchProps = Omit<BaseInputComponentProps<SwitchElementType>, keyof SwitchCustomProps | "as"> &
     SwitchCustomProps;
 
+// COMPONENT ///////////////////////////////////////////////////////////////////////////////////////////////////////////
 export const Switch = React.forwardRef(
     ({ size = "medium", ...props }: SwitchProps, ref: React.Ref<SwitchElementType>) => {
         return (