+++ CHANGELOG.md
@@ -678,6 +678,10 @@
 - TODO: Add mask design to PinInputField
 - TODO: Fix Sidebar scrolling issue
 - TODO: Revamp BreadCrumbs?
+- TODO: Add Tooltip component
+- TODO: Drag and drop file upload component
+- TODO: Date picker
+- TODO: Button size medium padding value
 
 ~~TODO: ⚠️ Fix TS issue in Select and SWS~~
 ~~TODO: ⚠️ Fix Meter height issue)~~